requirements.txt
PyYAML==6.0.2
requests==2.32.3
openai==1.40.3

